chap6-Pytorch实践-计算机视觉/6.4-InstanceSegmentation/requirements.txt
numpy==1.21.5
Pillow==9.2.0
pycocotools==2.0.5
torch==1.11.0
torchvision==0.12.0
